# Amortize allocs/pool calls in ReverseMatch/ForwardMatch by reusing iterator