This diff tweaks the active_ruleset and filter code paths to support reusing the *same* iterator across calls, instead of going to the pool for a new one every
time. In our service profiles for ReverseMatch, we saw a decent amount of contention from pool calls. In addition, previous code wasn't even closing the iterators it received, making pooling basically useless.

We need an iterator over the same ID multiple times within a single ReverseMatch/ForwardMatch call, so we can simply grab one at the top of the flow and reuse it throughout.

* filters code: has an API change to just accept an iterator directly
* rules code: no public facing API changes in this one.

Idea developed in conjunction with Prateek.

Benchstat comparison between old and new:

```
benchstat andrew_scratch/2024-10-08_outage/baseline andrew_scratch/2024-10-08_outage/amortize_allocs
goos: darwin
goarch: arm64
pkg: github.com/m3db/m3/src/metrics/rules
│ andrew_scratch/2024-10-08_outage/baseline │ andrew_scratch/2024-10-08_outage/amortize_allocs │
│                  sec/op                   │          sec/op           vs base                │
ActiveRuleSet_ReverseMatch/rollup_ID-10                                 52.73m ± ∞ ¹               47.69m ± ∞ ¹       ~ (p=1.000 n=1) ²
ActiveRuleSet_ReverseMatch/mapping_ID-10                                533.9m ± ∞ ¹               485.6m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                                                 167.8m                     152.2m        -9.30%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

│ andrew_scratch/2024-10-08_outage/baseline │ andrew_scratch/2024-10-08_outage/amortize_allocs │
│                   B/op                    │          B/op            vs base                 │
ActiveRuleSet_ReverseMatch/rollup_ID-10                                41.05Mi ± ∞ ¹             31.43Mi ± ∞ ¹        ~ (p=1.000 n=1) ²
ActiveRuleSet_ReverseMatch/mapping_ID-10                              1038.7Mi ± ∞ ¹             826.1Mi ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                206.5Mi                   161.1Mi        -21.96%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

│ andrew_scratch/2024-10-08_outage/baseline │ andrew_scratch/2024-10-08_outage/amortize_allocs │
│                 allocs/op                 │        allocs/op         vs base                 │
ActiveRuleSet_ReverseMatch/rollup_ID-10                                 360.0k ± ∞ ¹              270.0k ± ∞ ¹        ~ (p=1.000 n=1) ²
ActiveRuleSet_ReverseMatch/mapping_ID-10                                4.670M ± ∞ ¹              2.680M ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                 1.297M                    850.7k        -34.39%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

---

**Stack**:
- #4308 ⬅
- #4307


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do not merge manually using the UI - doing so may have unexpected results.*