@andrewmains12 2024-10-18
This approach breaks assumptions made by unchecked_metric_tags_iterator ; going to try an approach more in line with what we're doing for ForwardMatch.